Scale down iot-operator before beginning upgrade
* Split upgrader into separate module
* Run upgrader before starting controller loops
* Refactor scaling code into separate function.
* Stop iot-operator if it exists
* When scaling up, wait until readiness probe has passed
