+++ templates/csv/enmasse.clusterserviceversion.yaml
@@ -631,7 +631,7 @@ spec:
                   value: "true"
                 - name: CONTROLLER_ENABLE_AUTHENTICATION_SERVICE
                   value: "true"
-                - name: CONTROLLER_ENABLE_ADDRESS_SPACE_CONTROLLER_DEPLOYER
+                - name: CONTROLLER_ENABLE_ADDRESS_SPACE_CONTROLLER
                   value: "true"
                 - name: ADDRESS_SPACE_CONTROLLER_IMAGE
                   value: ${ADDRESS_SPACE_CONTROLLER_IMAGE}
